refactor | 기능 테스트 | FRB-252 | 테스트 수정 | 조수빈
